+++ exec/Chainweb/RichList.hs
@@ -1,109 +1,102 @@
-{-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE LambdaCase #-}
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
-module Chainweb.RichList
-( richList
-) where
+module Chainweb.RichList ( richList ) where
 
-
+import Control.Exception
 import Control.Monad
-
 import qualified Data.ByteString.Lazy as LBS
 import qualified Data.Csv as Csv
-import Data.Foldable (traverse_)
-import Data.List (sortOn, isPrefixOf, sort)
+import Data.Foldable (foldrM)
+import Data.List (isPrefixOf, sort,sortOn)
 import qualified Data.Map.Strict as M
-import           Data.Text (Text)
 import Data.Ord (Down(..))
-import qualified Data.Vector as V
+import qualified Data.Text as T
+import Data.Text (Text)
+import Data.String.Conv
+import Data.String
 
 import System.Directory
 import System.FilePath
-import System.Process
 import System.Logger.Types
-import Text.Read (readMaybe)
-import Text.Printf (printf)
+
+import Text.Read
+
+import Database.SQLite3
+import Database.SQLite3.Direct (Utf8(..))
+
+import Pact.Types.SQLite
 
 
 richList :: LogFunctionIO Text -> FilePath -> IO ()
 richList logger fp = do
-    --
-    -- Steps:
-    --   1. Check whether specified top-level db path is reachable
-    --   2. We assume the node data db is up to date, and for chains 0..19,
-    --      Check that the sqlite db paths exist. If yes, copy them to current
-    --      working dir.
-    --   3. Execute richlist generation, outputing `richlist.csv`
-    --   4. Aggregate richest accounts and prune to top 100
-    --
-    chains <- doesPathExist fp >>= \case
-      True -> copyTables
-      False -> ioError $ userError
-        $ "Chainweb-node top-level db directory does not exist: "
-        <> fp
 
-    logger Info $ "Aggregating richlist.csv..."
-    let cmd = proc "/bin/sh" ["scripts/richlist.sh", show chains]
-    void $! readCreateProcess cmd []
-
-    logger Info $ "Filtering top 100 richest accounts..."
-    void $! pruneRichList
-
-    logger Info $ "Finished."
+    files <- doesPathExist fp >>= \case
+        True -> checkChains
+        False -> ioError $ userError
+          $ "Chainweb-node top-level db directory does not exist: "
+          <> fp
+    logger Info "Aggregating richlist ..."
+    results <- foldrM (\file acc -> (acc .) <$> withSQLiteConnection file richListQuery) id files
+    logger Info $ "Filterting top 100 richest accounts..."
+    pruneRichList (results [])
   where
-    pruneRichList = do
-      csv <- LBS.readFile "richlist.csv"
-      case Csv.decode Csv.HasHeader csv of
-        Left e -> ioError $ userError $ "Could not decode rich list .csv file: " <> e
-        Right (rs :: V.Vector (String,String,String)) -> do
-          let go acc (acct,_,bal)
-                | bal == "balance" = acc
-                -- | otherwise = M.insertWith (+) acct (read @Double bal) acc
-                | otherwise = let err = error $ printf "richList: Couldn't read balance for account %s" acct
-                in M.insertWith (+) acct (maybe err id $ readMaybe @Double bal) acc
+    checkChains :: IO [FilePath]
+    checkChains = do
+        let sqlitePath = fp <> "chainweb-node/mainnet01/0/sqlite" -- TODO: notice this only works for mainnet
 
-          let acc = Csv.encode
-                $ take 100
-                $ sortOn (Down . snd)
-                $ M.toList
-                $ V.foldl' go M.empty rs
+        doesPathExist sqlitePath >>= \case
+          False -> ioError $ userError "Cannot find sqlite data. Is your node synced?"
+          True -> do
+            files <- filter ((==) ".sqlite" . takeExtension) <$> listDirectory sqlitePath
 
-          void $! LBS.writeFile "richlist.csv" acc
+            let go p
+                  | "pact-v1-chain-" `isPrefixOf` p =
+                    case splitAt 14 (fst $ splitExtension p) of
+                      (_, "") -> error $ "Found corrupt sqlite path: " <> p
+                      (_, cid) -> case readMaybe @Int cid of
+                        Just c -> ((p :), (c :))
+                        Nothing -> error "Couldn't read chain id"
+                  | otherwise = mempty
+                (fdl, cdl) = foldMap go files
+                chains = cdl []
+                isConsecutive = all (\(x,y) -> succ x == y)
+                  . (zip <*> tail)
+                  . sort
+            unless (isConsecutive chains)
+              $ ioError $ userError
+              $ "Missing tables for some chain ids. Is your node synced?"
+            return $ fdl []
 
-    copyTables :: IO Int
-    copyTables = do
-      let sqlitePath = fp </> "chainweb-node/mainnet01/0/sqlite"
+pruneRichList :: [(Text,Double)] -> IO ()
+pruneRichList = LBS.writeFile "richlist.csv"
+    . Csv.encode
+    . take 100
+    . sortOn (Down . snd)
+    . M.toList
+    . M.fromListWith (+)
 
-      doesPathExist sqlitePath >>= \case
-        False -> ioError $ userError $ "Cannot find sqlite data. Is your node synced?"
-        True -> do
-          dir <- filter ((==) ".sqlite" . takeExtension) <$> listDirectory sqlitePath
+-- WARNING: This function will throw errors if found. We don't "catch" errors in an Either type
+withSQLiteConnection :: FilePath -> (Database -> IO a) -> IO a
+withSQLiteConnection fp action = bracket (open (T.pack fp)) close action
 
-          -- count the number of sqlite files and aggregate associated file paths
-          --
-          let f (ns,acc) p
-                | "pact-v1-chain-" `isPrefixOf` p =
-                  -- this is not a magical 14 - this is the number of chars in "pact-v1-chain-"
-                  case splitAt 14 (fst $ splitExtension p) of
-                    (_, "") -> ioError $ userError $ "Found corrupt sqlite path: " <> p
-                    (_, cid) -> return ((read @Int cid):ns,p:acc)
-                | otherwise = return (ns,acc)
+type DList a = [a] -> [a]
 
-          (chains, files) <- foldM f mempty dir
+richListQuery :: Database -> IO (DList (Text, Double))
+richListQuery db = do
+    rows <- qry_ db (Utf8 richListQueryStmt) [RText, RInt, RDouble]
+    return $ (`foldMap` rows) $ \case
+      [SText (Utf8 account), SInt _txid, SDouble balance] -> ((toS account,balance) :)
+      _ -> error "impossible?" -- TODO: Make this use throwError/throwM instead of die
 
-          let isConsecutive = all (\(x,y) -> succ x == y)
-                . (zip <*> tail)
-                . sort
-
-          unless (isConsecutive chains)
-            $ ioError $ userError
-            $ "Missing tables for some chain ids. Is your node synced?"
-
-          -- copy all files to current working dir
-          traverse_ (\p -> copyFile (sqlitePath </> p) p) files
-
-          -- return # of chains for bash
-          return $ length chains
+richListQueryStmt :: IsString s => s
+richListQueryStmt =
+   "select rowkey as acct_id, txid, cast(ifnull(json_extract(rowdata, '$.balance.decimal'), json_extract(rowdata, '$.balance')) as REAL) as 'balance'\
+     \ from [coin_coin-table] as coin\
+     \ INNER JOIN (\
+      \ select\
+       \ rowkey as acct_id,\
+       \ max(txid) as last_txid\
+      \ from 'coin_coin-table'\
+      \ group by acct_id\
+     \ ) latest ON coin.rowkey = latest.acct_id AND coin.txid = latest.last_txid\
+     \ order by balance desc;"
